Expose the responses in our Honey class.
This means keeping around a blocking queue for
containing the responses (or the events overflow
errors).

A thing to notice is that, upon disposal, we will try
to add a null value to signal the user we are done -
but if the responses queue is full, we will ignore it.
This may happen due mostly the user ignoring to consume
the responses queue, having it reach its full capacity.